Refactor frontend structure for Restaurants Page
This commit introduces a series of changes aimed at improving the structure of the frontend. The idea was to encapsulate the Restaurants Page inside its own component and use a page template for consistent design across the application. This provides easier maintenance and scalability. Furthermore, a Header component and Logo component were added as part of the new template. Besides, more entries were added to the restaurants data for testing purposes.