Add new restaurant functionality
Added a new functionality to add a restaurant. Created a NewRestaurantDTO to handle the incoming data. The RestaurantController now includes a POST mapping to add a new restaurant using this DTO. The RestaurantService has also been updated to include an addRestaurant method, which saves the new restaurant to the repository.