Add logging for restaurant actions
Incorporated the use of logtail for tracking actions and potential errors related to restaurant data. Logging was added to essential actions such as receiving, updating, and creating new restaurant data in RestaurantDetailsPage, RestaurantEditPage, AddRestaurantsPage, and RestaurantsPage. This enables us to better understand the flow of data and promptly identify errors or issues.