Add comprehensive project information to README
The README file has been updated extensively to include more detailed information about the project. The added information includes an introduction, prerequisites, dependencies, a getting started guide, how to run tests, and a list of authors. This enhancement will make the project easier to understand and get started with for new developers.