Add user data loading to ProfilePage
The UserType model has been introduced and a user data loading function has been added to the ProfilePage component. The function utilizes axios to make the request to the "/api/auth/me" endpoint and sets the user data on successful response. The login functionality was moved from the Header component to the ProfilePage component.